[tasks] revamp anilist tasks to not be a memory hog
Currently we make huge requests to anilist with all the data that we
want, and merge the new data when we get to it. It uses a lot of memory
when running (several GBs at this point) and it will keep growing as we
go.
So this revamp removes the multiton and other requests abstractions that
get in our way and don’t make much sense (we don’t need an in-memory
cache, we can and do query the database for most of our needs).

The requests are much lighter, requesting item ids instead of the whole
object. we keep the last update time to choose which items we want to
update in priority (at this point I’m using arbitrary, but most likely
fine, numbers to choose them).

The first sync is certainly slower, but not in a ridiculous way. with my
own anilist account in the db I expect it to take a few hours.
The memory usage seems to stay around 200-300M throughout the run.
When coming from a previous db all items will be requested to set the
update time to a non-zero value which means that the first sync after
this change will be also a bit longer.

Some items might be added by running the script several time (without
changes on AniList’s end), and I would expect the task to run every day
or even several times per day.
Starting from a fresh database the medias that are not in the user’s
list might not be linked to their characters on the first run of the
script which should fix itself on a second run (which might find more
orphan medias that would require another run to be linked to their
medias), some mitigations are possible but I think this is unavoidable
and in practice won’t be a problem.

It should be easy to crawl through some other medias periodically (new,
current season, random) to feed the database more often and holistically
than we were doing previously.

Probably should try starting from a fresh database to see
if it is tolerable.